add check that user is logged in

+++ src/main/java/com/andreas/server/controller/FileServerController.java
@@ -2,11 +2,13 @@ package com.andreas.server.controller;
 
 import com.andreas.common.FileMetaDTO;
 import com.andreas.common.FileServer;
+import com.andreas.common.NotLoggedInException;
 import com.andreas.common.UserDTO;
 import com.andreas.server.database.DatabaseException;
 import com.andreas.server.database.FileServerDAO;
 import com.andreas.server.database.FileServerDAOImpl;
 import com.andreas.server.model.FileMetaData;
+import com.andreas.server.model.LoginManager;
 import com.andreas.server.model.User;
 
 import java.rmi.RemoteException;
@@ -17,12 +19,15 @@ import java.util.List;
 public class FileServerController extends UnicastRemoteObject implements FileServer{
 
     private final FileServerDAO fileServerDAO;
+    private final LoginManager loginManager = new LoginManager();
 
     public FileServerController() throws DatabaseException, RemoteException {
         fileServerDAO = new FileServerDAOImpl();
     }
 
     public UserDTO login(String username, String password) throws DatabaseException, RemoteException {
+        UserDTO user = fileServerDAO.login(username, password);
+        loginManager.addLoggedInUser(user);
         return fileServerDAO.login(username, password);
     }
 
@@ -33,28 +38,38 @@ public class FileServerController extends UnicastRemoteObject implements FileSer
             if (userDTO.getName().equals(username))
                 return null;
         }
-        return fileServerDAO.insertUser(username, password);
+        UserDTO user = fileServerDAO.insertUser(username, password);
+        loginManager.addLoggedInUser(user);
+        return user;
     }
 
     @Override
-    public void logout(UserDTO user) {
-        //TODO: implement some kind of check
+    public void logout(UserDTO user) throws NotLoggedInException {
+        if (loginManager.isLoggedIn(user))
+            throw new NotLoggedInException();
+        loginManager.removeLoggedInUser(user);
     }
 
     @Override
-    public FileMetaDTO uploadFile(String filename, UserDTO currentUser, boolean readOnly, boolean publicAccess) throws RemoteException, DatabaseException {
+    public FileMetaDTO uploadFile(String filename, UserDTO currentUser, boolean readOnly, boolean publicAccess) throws RemoteException, DatabaseException, NotLoggedInException {
+        if (!loginManager.isLoggedIn(currentUser))
+            throw new NotLoggedInException();
         FileMetaData fileMetaData = new FileMetaData(filename, new User(currentUser.getId(), currentUser.getName()), readOnly, publicAccess);
         fileServerDAO.insertFile(fileMetaData);
         return fileMetaData;
     }
 
     @Override
-    public List<FileMetaDTO> getFiles(UserDTO user) throws DatabaseException {
+    public List<FileMetaDTO> getFiles(UserDTO user) throws DatabaseException, NotLoggedInException {
+        if (!loginManager.isLoggedIn(user))
+            throw new NotLoggedInException();
         return fileServerDAO.getFiles(user);
     }
 
     @Override
-    public void unregister(UserDTO currentUser) throws DatabaseException {
+    public void unregister(UserDTO currentUser) throws DatabaseException, NotLoggedInException {
+        if (!loginManager.isLoggedIn(currentUser))
+            throw new NotLoggedInException();
         fileServerDAO.removeUser(currentUser);
     }
